move code to vmSlice
